プレイヤー脱落後の処理を作成。他
プレイヤーが死亡した後にカメラを自由に動かせるようにした。
ロケットを渡されたときに、一定時間動けなくなるようになった。
