+++ Rocket_Tag/Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -44,6 +44,8 @@ MonoBehaviour:
   - ChangeColor
   - AddPlayerList
   - StartGame
+  - SetPlayerDead
+  - SetIsStun
   DisableAutoOpenWizard: 1
   ShowSettings: 0
   DevRegionSetOnce: 1
