+++ Rocket_Tag/Assets/Resources/Materials/MapMaterial/Map_01_Color.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9fa175475fd76c4e9a8c044a97882b6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
