+++ Rocket_Tag/Assets/Resources/Materials/MapMaterial/Map_02_Color.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae8772ecab0048048bf873ae7d30684b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
