+++ Rocket_Tag/Assets/Resources/Materials/MapMaterial/Map_03_Color.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b633aba839aa5a743881837593c532c5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
