+++ Rocket_Tag/Assets/Resources/Materials/MapMaterial/Map_04_Color.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6dceb5c54a29f0a4fbf1a3a3b4f5be49
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
