+++ Rocket_Tag/Assets/Resources/Materials/PlayerMaterial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 048fdc61eb3ed2d489ca044c3a4eee0f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
