+++ Rocket_Tag/Assets/Resources/Materials/PlayerMaterial/Player_StunBodyColor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1c574ffc29483364abdf102a1c5a1ce2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
