+++ Rocket_Tag/Assets/Resources/Materials/RocketMaterial/Rocket_DangerColor.mat.meta
