--- Rocket_Tag/Assets/Scripts/DebugButton.cs
@@ -1,10 +1,0 @@
-using Photon.Pun;
-using UnityEngine;
-
-public class DebugButton : MonoBehaviourPunCallbacks
-{
-    public void OnClick()
-    {
-        
-    }
-}
