--- Rocket_Tag/Assets/Scripts/DebugButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2ea300aa9d7c4954cb022604019d5b7b
